standardize function names again
